Implement multidimensional indexing
This allows the subscript to AtIndex/AssignAtIndex to be a list,
thus multidimensionally indexing the source array.
Futhermore, each list entry may itself be a list,
allowing multiple indices to be retrieved or stored at once.
In the latter case, a list of values can be specified,
which is itself consumed cyclically.
It can also be used to apply a permutation to a list,
which is convenient if the list is computed inline.