Implement transpose called Zip
- Zip(number) "transposes" its real and complex parts
- Zip(string) "transposes" it to a list of characters
- Zip([string]) transposes it to a list of strings
- Zip([[any]]) transposes the first two dimensions of the array

If the array has nonincreasing width then the transpose will be reversible:
AtIndex(AtIndex(Zip(u), l), k) returns the same as AtIndex(AtIndex(u, k), l).

Otherwise the gaps in the lists or strings will be closed up e.g.
Zip(['this', 'is', 'a', 'test']) results in ['tiat', 'hse', 'is', 'st'].

Zip will also try to work on dicts. A list of dicts may not be reversible
but a dict of nonempty lists or dicts is always reversible.

If the elements of the list have mixed type then the worst case type is used.